Makes class non-anonymous.
Jira: OTPL-1774
